Switch to using fetchzip instead of fetchurl
GitHub's source tarballs are created in a non-deterministic way and the
order the entries are added is not stable.

In nixpkgs, there is fetchFromGitHub which works around this issue by
being a wrapper around fetchzip and since we're already using the full
URL to the corresponding GitHub archives, I switched to fetchzip instead
to keep the changes minimal.

Signed-off-by: aszlig <aszlig@nix.build>