test(*): automatically check all binaries, based on binNames
The test list would go out-of-date when new packages were added. This
does the same check as before, but uses the binaries we already
specify on the nix level (via a passthru).

The next step is to try to run the binaries (e.g. with --version),
which will give us a better test of whether they are actually
executable.